move Condition to value
