+++ cirq-core/cirq/ops/raw_types.py
@@ -423,35 +423,6 @@ class Gate(metaclass=value.ABCMetaImplementAnyOneOf):
         return protocols.obj_to_dict_helper(self, attribute_names=[])
 
 
-class Condition:
-    def __init__(self, expr: sympy.Expr, keys: Tuple[value.MeasurementKey, ...]):
-        self._expr = expr
-        self._keys = keys
-
-    @property
-    def keys(self):
-        return self._keys
-
-    @property
-    def expr(self):
-        return self._expr
-
-    def with_keys(self, keys: Tuple[value.MeasurementKey, ...]):
-        assert len(keys) == len(self._keys)
-        return Condition(self._expr, keys)
-
-    def __eq__(self, x):
-        return isinstance(x, Condition) and self._keys == x._keys and self._expr == x._expr
-
-    def __hash__(self):
-        return hash(self._keys) ^ hash(self._expr)
-
-    def __str__(self):
-        if self._expr == sympy.symbols('x0') and len(self._keys) == 1:
-            return str(self._keys[0])
-        return f'({self._expr}, {self._keys})'
-
-
 TSelf = TypeVar('TSelf', bound='Operation')
 
 
@@ -622,11 +593,11 @@ class Operation(metaclass=abc.ABCMeta):
         return np.allclose(m12, m21, atol=atol)
 
     @property
-    def classical_controls(self) -> FrozenSet[Condition]:
+    def classical_controls(self) -> FrozenSet['cirq.Condition']:
         return frozenset()
 
     def with_classical_controls(
-        self, *conditions: Union[str, 'cirq.MeasurementKey', Condition]
+        self, *conditions: Union[str, 'cirq.MeasurementKey', 'cirq.Condition']
     ) -> 'cirq.ClassicallyControlledOperation':
         """Returns a classically controlled version of this operation.
 
@@ -857,7 +828,7 @@ class TaggedOperation(Operation):
         return protocols.equal_up_to_global_phase(self.sub_operation, other, atol=atol)
 
     @property
-    def classical_controls(self) -> FrozenSet[Condition]:
+    def classical_controls(self) -> FrozenSet['cirq.Condition']:
         return self.sub_operation.classical_controls
 
     def without_classical_controls(self) -> 'cirq.Operation':
